Fix vote percentages
We work out averaege percent as:
sum(votes in each OA) / count(OAs)

But count(OAs) was group by (revised, party)

So count(OAs) was the count of OAs for which the *party* stood.
This would inflate the votes of any candidate which stood in only
one ward!

Also -- it's really slow to run it now. So we do one constituency
at a time.